Working on orders, user
made a user registration, so you have to register, and on checkout if you don't have an account it will redirect you to register. also fixed cartentities, instead of deleting them it will change the session id so you don't see it anymore. because if deleted the items will not show on the orders.